# [docker] update chisel build method to what is recommmended