--- a/.github/workflows/ci.yml
+++ b/.github/workflows/ci.yml
@@ -9,7 +9,6 @@ on:
     branches: ["main"]
   pull_request:
 jobs:
-
   ########
   # Docs #
   ########
@@ -36,7 +35,7 @@ jobs:
     steps:
       - uses: actions/checkout@v2
         with:
-          submodules: 'recursive'
+          submodules: "recursive"
       - name: Install DNN Python Modules
         run: |
           pip3 install -r sw/dnn/requirements.txt
@@ -55,7 +54,6 @@ jobs:
           sw/standard-fp.yaml sw/openmp.yaml sw/snitch-cluster-openmp.yaml \
           sw/blas.yaml sw/dnn.yaml -j
 
-
   ##############################################
   # Simulate SW on SNAX Cluster w/ Verilator #
   ##############################################
@@ -68,7 +66,7 @@ jobs:
     steps:
       - uses: actions/checkout@v2
         with:
-          submodules: 'recursive'
+          submodules: "recursive"
       - name: Build Software
         run: |
           make -C target/snitch_cluster sw \
@@ -93,7 +91,7 @@ jobs:
     steps:
       - uses: actions/checkout@v2
         with:
-          submodules: 'recursive'
+          submodules: "recursive"
       - name: Build Software
         run: |
           make -C target/snitch_cluster sw \
@@ -118,7 +116,7 @@ jobs:
     steps:
       - uses: actions/checkout@v2
         with:
-          submodules: 'recursive'
+          submodules: "recursive"
       - name: Build Software
         run: |
           make -C target/snitch_cluster sw \
@@ -143,7 +141,7 @@ jobs:
     steps:
       - uses: actions/checkout@v2
         with:
-          submodules: 'recursive'
+          submodules: "recursive"
       - name: Build Software
         run: |
           make -C target/snitch_cluster sw \
@@ -165,7 +163,7 @@ jobs:
     name: Simulate SW on SNAX ALU Cluster w/ Verilator (Generic LLVM)
     runs-on: ubuntu-22.04
     container:
-      image: ghcr.io/kuleuven-micas/snax:main
+      image: ghcr.io/kuleuven-micas/snax:pr-110
     steps:
       - uses: actions/checkout@v2
         with:
@@ -199,7 +197,7 @@ jobs:
     steps:
       - uses: actions/checkout@v2
         with:
-          submodules: 'recursive'
+          submodules: "recursive"
       - name: Install DNN Python Modules
         run: |
           pip3 install -r sw/dnn/requirements.txt