--- a/hw/chisel/build.sbt
+++ b/hw/chisel/build.sbt
@@ -4,7 +4,7 @@ ThisBuild / scalaVersion := "2.13.8"
 ThisBuild / version := "0.1.0"
 ThisBuild / organization := "be.kuleuven.esat.micas"
 
-val chiselVersion = "5.0.0"
+val chiselVersion = "6.3.0"
 
 lazy val root = (project in file("."))
   .settings(