--- a/util/wrappergen/wrappergen.py
+++ b/util/wrappergen/wrappergen.py
@@ -56,7 +56,6 @@ def gen_chisel_file(chisel_path, chisel_param, gen_path):
     # Call chisel environment and generate the system verilog file
     cmd = f" cd {chisel_path} && \
         mill Snax.runMain {chisel_param} {gen_path}"
-    print(cmd)
     os.system(cmd)
 
     return