The only thing that wasn't updated in #96 was the Chisel version.
This PR changes the way we build Chisel stuff to be up to date with what is recommended by the developers of Chisel.
( sorry for all the auto-formatting changes, we should standardize autoformatting for everything sometime)

## Problems

We currently have the following problems:
- We are not using the most recent Chisel version (6.3)
- We are not using `scala-cli`
[link](https://www.chisel-lang.org/docs/installation)
> As described above, Scala CLI is a great "batteries included" way to use Chisel. It will automatically download and manage all dependencies of Chisel including a Java Development Kit (JDK). More complex projects will require the user to install a JDK and a build tool.
With `scala-cli`, we don't have to manage everything manually, which is very nice
- We are using `sbt`
[link](https://www.chisel-lang.org/docs/installation)
> [Mill](https://mill-build.com/) is a modern Scala build tool with simple syntax and a better command-line experience than SBT. We recommend Chisel users use Mill.

Other people have quite [strong opinions](https://www.lihaoyi.com/post/SowhatswrongwithSBT.html) about sbt:
> sbt's syntax is just honestly sinful. It looks like something the germans would have come up with... 70 yrs ago... in a submarine.

Me personally, I have also ran into quite annoying problems with sbt, with no docs to solve these problems, so I hope mill is a better alternative

- We are using `ChiselTest`
[link](https://www.chisel-lang.org/docs/appendix/migrating-from-chiseltest)
> ChiselTest is not used or maintained by the core Chisel development team or their employers. ChiselSim is the approved replacement for ChiselTest in Chisel 5 and beyond. ChiselSim is maintained and used by the core Chisel development team. This page describes how to migrate from ChiselTest to ChiselSim.

However the state of ChiselSim seems to be even worse:
https://github.com/chipsalliance/chisel/discussions/3957
https://github.com/chipsalliance/chisel/discussions/4033

And berkeley seems to still put effort in ChiselTest, so let's keep using that.

## Solutions

Therefore the PR employs the following changes:

- use ChiselVersion 6.3 with ChiselTest 6.0
- add a build.sc to support builds with mill
- add scala-cli and mill to the docker build
- remove firttl from the docker build
- run CI using mill

## Results

This has the following advantages:

- automatic version resolution of the firttl compiler (no need to check correct version ourself)
- chisel 6 features (some new warnings are generated from things we do wrong :) )
- much faster builds with mill:
  - formatting: 8s -> 0.5s
  - compiling (no cache): 20s -> 11s
  - compiling (with cache): 10s -> 0.8s

## mill user guide

I kindly ask to test out the mill build tool instead of `sbt`, it results in much faster builds. Here is a short user guide to help with this conversion:

instead of `sbt runMain`, use:
```
mill Snax.runMain
```

instead of `sbt test`, use:
```
mill Snax.test
```

instead of `sbt testOnly`, use:
```
mill Snax.test.testOnly "snax.csr_manager.CsrManagerTest"
```

Formatting:
```
mill Snax.reformat
mill Snax.checkFormat
```

Check other commands with either
```
mill resolve Snax._
mill --help
```

## Todos

- [ ] after approval, reset docker tags to main