@xiaoling-yi 2024-05-17
Hi @jorendumoulin  @rgantonio , I am fine with this PR. Personally, I didn't encounter any problem using `sbt`. If you decide to change from `sbt` to `mill` in the SNAX cluster. Please make sure to change the commands in the related Makefile accordingly.