@JosseVanDelm 2024-05-20
Hi! I'm definitely for the integration of this new stuff.

The build times are indeed much faster, even though the main bottleneck is still simulation
I also like to keep open possibilities in automatic hardware generation with chisel-related flows, so doing a periodic update every once in a while seems like a great idea (especially before we do the tape-out and we still have some freedom in changing these things).

> Please make sure to change the commands in the related Makefile accordingly.

@xiaoling-yi it seems that all invocations of (previously) sbt are now inside the python scripts, and I think @jorendumoulin updated those all correctly. Could you confirm this is the case? Or if we are still missing something? Thanks!

> chisel 6 features (some new warnings are generated from things we do wrong :) )

@jorendumoulin what kind of warnings are generated?
I'm wondering if the changes proposed in this PR are transparent for the user, otherwise we might have to warn people during the tutorial that there are some changes incoming, unless we manage to integrate this before the tutorial.
@xiaoling-yi @rgantonio we might want to rebase this as you suggested and give priority to the other stuff needed for the tutorial, right? Apart from #111 are there other things we should wait for here?