@xiaoling-yi 2024-05-20
> > Please make sure to change the commands in the related Makefile accordingly.
> 
> @xiaoling-yi it seems that all invocations of (previously) sbt are now inside the python scripts, and I think @jorendumoulin updated those all correctly. Could you confirm this is the case? Or if we are still missing something? Thanks!

Hi @JosseVanDelm , yes, the `sbt` cmd is in python script now and the `sbt` build possibility has remained so I think it is fine!

> > chisel 6 features (some new warnings are generated from things we do wrong :) )
> 
> @jorendumoulin what kind of warnings are generated? I'm wondering if the changes proposed in this PR are transparent for the user, otherwise we might have to warn people during the tutorial that there are some changes incoming, unless we manage to integrate this before the tutorial. @xiaoling-yi @rgantonio we might want to rebase this as you suggested and give priority to the other stuff needed for the tutorial, right? Apart from #111 are there other things we should wait for here?

As for merging before or after the tutorial, both are ok for me! @rgantonio what do you think as you a in main charge of the tutorial?