@rgantonio 2024-05-20
@xiaoling-yi since `sbt` is still working anyway, we can push this in and update the scripts later. We're not mentioning the use of `sbt` or `mill` anyway in the tutorial so it's okay 😄