@jorendumoulin 2024-05-20
Okay, thank you for all the comments! 
Yes, `sbt` is still fully functional, it seems common practice to support both build systems. One example: https://github.com/riscv-boom/riscv-boom. 
The main thing here is the upgrading of the chisel version. The development of the chisel language and tools is moving fast and has cool prospects, so to me it seems useful to keep updating the version along, definitely for core framework components.

> @jorendumoulin what kind of warnings are generated?

They are concerning non-matching signal widths in the Csr Manager. No problem for simulation, but as this seems to have caused some issues for synthesis, it could be nice to clean this up in the future :).

Ok! I rebased to main, I will merge! (The latest build is failing because it now runs again on the main docker branch, this will be fixed as soon as the new main container image is built.)